requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
colorama==0.4.6
Django==4.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
flake8==6.0.0
iniconfig==1.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==22.0
parameterized==0.8.1
Pillow==9.3.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
